api/requirements.txt
mysql-connector-python==8.0.31
pandas==1.3.5
Flask==2.2.2
waitress==2.1.2